requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
pyodbc==5.0.1
python-dotenv==1.0.1
Werkzeug==3.0.1
pytest==8.0.0
pytest-cov
black